SLING-10477 : Review usage of UserEntryHandler in RecollectorVaultPackageScanner (tests)
