+++ src/test/java/org/apache/sling/feature/cpconverter/ContentPackage2FeatureModelConverterTest.java
@@ -81,8 +81,7 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 
-@RunWith(Parameterized.class)
-public class ContentPackage2FeatureModelConverterTest {
+public class ContentPackage2FeatureModelConverterTest extends AbstractConverterTest {
 
     /**
      * Test package A-1.0. Depends on B and C-1.X
@@ -100,23 +99,10 @@ public class ContentPackage2FeatureModelConverterTest {
 
     private ContentPackage2FeatureModelConverter converter;
     private EntryHandlersManager handlersManager;
-    
-    private final String systemUserRelPath;
-    
-    @Parameterized.Parameters(name = "name={1}")
-    public static Collection<Object[]> parameters() {
-        return Arrays.asList(
-                new Object[] {ConverterConstants.SYSTEM_USER_REL_PATH_DEFAULT, "Default system user rel-path"},
-                new Object[] { "system/cq:services", "Modified system user rel-path"});
-    }
-    
-    public ContentPackage2FeatureModelConverterTest(@NotNull String systemUserRelPath, @NotNull String name) {
-        this.systemUserRelPath = systemUserRelPath;
-    }
         
     @Before
     public void setUp() {
-        handlersManager = new DefaultEntryHandlersManager(Collections.emptyMap(), false, ContentPackage2FeatureModelConverter.SlingInitialContentPolicy.KEEP, systemUserRelPath);
+        handlersManager = new DefaultEntryHandlersManager();
         converter = new ContentPackage2FeatureModelConverter()
                     .setEntryHandlersManager(handlersManager)
                     .setAclManager(new DefaultAclManager());
@@ -511,129 +497,6 @@ public class ContentPackage2FeatureModelConverterTest {
         }
     }
 
-    private static void deleteDirTree(File dir) throws IOException {
-        Path tempDir = dir.toPath();
-
-        Files.walk(tempDir)
-            .sorted(Comparator.reverseOrder())
-            .map(Path::toFile)
-            .forEach(File::delete);
-    }
-
-    private static Feature getFeature(File outputDirectory, String name) throws Exception {
-        File featureFile = new File(outputDirectory, name);
-        assertTrue(featureFile + " was not correctly created", featureFile.exists());
-
-        try (Reader reader = new FileReader(featureFile)) {
-            return FeatureJSONReader.read(reader, featureFile.getAbsolutePath());
-        }
-    }
-
-    private static void verifyFeatureFile(File outputDirectory,
-                                   String name,
-                                   String expectedArtifactId,
-                                   List<String> expectedArtifacts,
-                                   List<String> expectedConfigurations,
-                                   List<String> expectedContentPackagesExtensions) throws Exception {
-        Feature feature = getFeature(outputDirectory, name);
-
-        assertEquals(expectedArtifactId, feature.getId().toMvnId());
-
-        for (String expectedArtifact : expectedArtifacts) {
-            assertTrue(expectedArtifact + " not found in Feature " + expectedArtifactId, feature.getBundles().containsExact(ArtifactId.fromMvnId(expectedArtifact)));
-            verifyInstalledArtifact(outputDirectory, expectedArtifact);
-        }
-
-        for (String expectedConfiguration : expectedConfigurations) {
-            assertNotNull(expectedConfiguration + " not found in Feature " + expectedArtifactId, feature.getConfigurations().getConfiguration(expectedConfiguration));
-        }
-
-        if (expectedContentPackagesExtensions.size() > 0) {
-        Artifacts contentPackages = feature.getExtensions().getByName("content-packages").getArtifacts();
-        assertEquals(expectedContentPackagesExtensions.size(), contentPackages.size());
-
-            for (String expectedContentPackagesExtension : expectedContentPackagesExtensions) {
-                assertTrue(expectedContentPackagesExtension + " not found in Feature " + expectedArtifactId,
-                    contentPackages.containsExact(ArtifactId.fromMvnId(expectedContentPackagesExtension)));
-                verifyInstalledArtifact(outputDirectory, expectedContentPackagesExtension);
-            }
-        }
-    }
-
-    private static void verifyInstalledArtifact(File outputDirectory, String coordinates) {
-        ArtifactId bundleId = ArtifactId.fromMvnId(coordinates);
-
-        StringTokenizer tokenizer = new StringTokenizer(bundleId.getGroupId(), ".");
-        while (tokenizer.hasMoreTokens()) {
-            outputDirectory = new File(outputDirectory, tokenizer.nextToken());
-        }
-
-        outputDirectory = new File(outputDirectory, bundleId.getArtifactId());
-        outputDirectory = new File(outputDirectory, bundleId.getVersion());
-
-        StringBuilder bundleFileName = new StringBuilder()
-                                       .append(bundleId.getArtifactId())
-                                       .append('-')
-                                       .append(bundleId.getVersion());
-        if (bundleId.getClassifier() != null) {
-            bundleFileName.append('-').append(bundleId.getClassifier());
-        }
-        bundleFileName.append('.').append(bundleId.getType());
-
-        File bundleFile = new File(outputDirectory, bundleFileName.toString());
-        assertTrue("Bundle " + bundleFile + " does not exist", bundleFile.exists());
-
-        File pomFile = new File(outputDirectory, String.format("%s-%s.pom", bundleId.getArtifactId(), bundleId.getVersion()));
-        assertTrue("POM file " + pomFile + " does not exist", pomFile.exists());
-    }
-
-    private static void verifyContentPackageEntryNames(File contentPackage, String...expectedEntryNames) throws Exception {
-        try (ZipFile zipFile = new ZipFile(contentPackage)) {
-            List<String> expectedEntryNamesList = Arrays.asList(expectedEntryNames);
-            for (Enumeration<? extends ZipEntry> zipEntries = zipFile.entries(); zipEntries.hasMoreElements();) {
-                ZipEntry zipEntry = zipEntries.nextElement();
-                String entryName = zipEntry.getName();
-                if (!entryName.endsWith("/")) {
-                    assertTrue("ZipEntry with name " + entryName + " not expected", expectedEntryNamesList.contains(entryName));
-                }
-            }
-        }
-    }
-
-    private static void verifyContentPackage(File contentPackage, String...expectedEntries) throws Exception {
-        verifyContentPackage(contentPackage, Collections.emptySet(), expectedEntries);
-    }
-    
-    private static void verifyContentPackage(File contentPackage, @NotNull Set<String> notExpected, String...expectedEntries) throws Exception {
-        try (ZipFile zipFile = new ZipFile(contentPackage)) {
-            for (String expectedEntry : expectedEntries) {
-                assertNotNull("Expected entry not found: " + expectedEntry + " in file " + contentPackage, zipFile.getEntry(expectedEntry));
-            }
-            for (String notExpectedEntry : notExpected) {
-                assertNull("Not expected entry found: " + notExpectedEntry + " in file " + contentPackage, zipFile.getEntry(notExpectedEntry));
-            }
-        }
-    }
-
-    private static void verifyPropertiesXmlEntry(File contentPackage, String... expectedPropertyKeys) throws IOException {
-        try (ZipFile zipFile = new ZipFile(contentPackage)) {
-            ZipEntry zipEntry = zipFile.getEntry(Constants.META_DIR + "/" + Constants.PROPERTIES_XML);
-            assertNotNull("Package didn't contain properties.xml", zipEntry);
-            try (InputStream inputStream = zipFile.getInputStream(zipEntry)) {
-                Properties properties = new Properties();
-                properties.loadFromXML(inputStream);
-                for (String expectedPropertyKey : expectedPropertyKeys) {
-                    if (expectedPropertyKey.startsWith("!")) {
-                        String key = expectedPropertyKey.substring(1);
-                        assertFalse("Properties.xml was not supposed to contain key " +  key + " but it does", properties.containsKey(key));
-                    } else {
-                        assertTrue("Properties.xml was supposed to contain key " +  expectedPropertyKey + " but it does not", properties.containsKey(expectedPropertyKey));
-                    }
-                }
-            }
-        }
-    }
-
     @Test(expected = IllegalArgumentException.class)
     public void verifyFilteringOutUndesiredPackages() throws Exception {
         RegexBasedResourceFilter resourceFilter = new RegexBasedResourceFilter();
@@ -1000,166 +863,4 @@ public class ContentPackage2FeatureModelConverterTest {
 
         return loadedResources;
     }
-
-    @Test
-    public void testConvertPackageWithUsersGroupsAndServiceUsers() throws Exception {
-        URL packageUrl = getClass().getResource("demo-cp.zip");
-        File packageFile = FileUtils.toFile(packageUrl);
-        File outputDirectory = new File(System.getProperty("java.io.tmpdir"), getClass().getName() + '_' + System.currentTimeMillis());
-        try {
-            converter.setFeaturesManager(new DefaultFeaturesManager(true, 5, outputDirectory, null, null, null, new DefaultAclManager()))
-                    .setBundlesDeployer(new LocalMavenRepositoryArtifactsDeployer(outputDirectory))
-                    .setEmitter(DefaultPackagesEventsEmitter.open(outputDirectory))
-                    .convert(packageFile);
-
-            File converted = new File(outputDirectory, "my_packages/demo-cp/0.0.0/demo-cp-0.0.0-cp2fm-converted.zip");
-            Set<String> notExpected = new HashSet<>();
-            notExpected.add("jcr_root/apps/demo-cp/.content.xml");
-            notExpected.add("jcr_root/home/users/demo-cp/_rep_policy.xml");
-            notExpected.add("jcr_root/home/groups/demo-cp/_rep_policy.xml");
-            notExpected.add("jcr_root/home/users/system/.content.xml");
-            notExpected.add("jcr_root/home/users/system/_cq_services/.content.xml");
-            notExpected.add("jcr_root/home/users/system/_cq_services/demo-cp/.content.xml");
-            notExpected.add("jcr_root/home/users/system/_cq_services/demo-cp/qStDu7IQBLa95gURmer1/.content.xml");
-            notExpected.add("jcr_root/home/users/system/_cq_services/demo-cp/qStDu7IQBLa95gURmer1/_rep_principalPolicy.xml");
-            verifyContentPackage(converted, 
-                    notExpected,
-                    "META-INF/vault/properties.xml",
-                    "META-INF/vault/config.xml",
-                    "META-INF/vault/filter.xml",
-                    "jcr_root/.content.xml",
-                    "jcr_root/demo-cp/.content.xml",
-                    "jcr_root/demo-cp/_rep_policy.xml",
-                    "jcr_root/apps/.content.xml",
-                    "jcr_root/home/.content.xml",
-                    "jcr_root/home/users/demo-cp/.content.xml",
-                    "jcr_root/home/users/demo-cp/XPXhA_RKMFRKNO8ViIhn/.content.xml",
-                    "jcr_root/home/users/demo-cp/XPXhA_RKMFRKNO8ViIhn/_rep_policy.xml",
-                    "jcr_root/home/groups/.content.xml",
-                    "jcr_root/home/groups/demo-cp/.content.xml",
-                    "jcr_root/home/groups/demo-cp/EsYrXeBdSRkna2kqbxjl/.content.xml",
-                    "jcr_root/home/groups/demo-cp/EsYrXeBdSRkna2kqbxjl/_rep_policy.xml"
-                    );
-
-            assertPolicy(converted, "jcr_root/demo-cp/_rep_policy.xml", "cp-serviceuser-1", "cp-user1", "cp-group1");
-            assertPolicy(converted, "jcr_root/home/groups/demo-cp/EsYrXeBdSRkna2kqbxjl/_rep_policy.xml", null, "cp-group1");
-            assertPolicy(converted,  "jcr_root/home/users/demo-cp/XPXhA_RKMFRKNO8ViIhn/_rep_policy.xml", null, "cp-user1");
-        } finally {
-            deleteDirTree(outputDirectory);
-        }
-    }
-    @Test
-    public void testConvertPackageWithUsersGroupsAndServiceUsersRepPolicyFirst() throws Exception {
-        URL packageUrl = getClass().getResource("demo-cp3.zip");
-        File packageFile = FileUtils.toFile(packageUrl);
-        File outputDirectory = new File(System.getProperty("java.io.tmpdir"), getClass().getName() + '_' + System.currentTimeMillis());
-        try {
-            converter.setFeaturesManager(new DefaultFeaturesManager(true, 5, outputDirectory, null, null, null, new DefaultAclManager()))
-                    .setBundlesDeployer(new LocalMavenRepositoryArtifactsDeployer(outputDirectory))
-                    .setEmitter(DefaultPackagesEventsEmitter.open(outputDirectory))
-                    .convert(packageFile);
-
-            File converted = new File(outputDirectory, "my_packages/demo-cp/0.0.0/demo-cp-0.0.0-cp2fm-converted.zip");
-            Set<String> notExpected = new HashSet<>();
-            notExpected.add("jcr_root/apps/demo-cp/.content.xml");
-            notExpected.add("jcr_root/home/users/demo-cp/_rep_policy.xml");
-            notExpected.add("jcr_root/home/groups/demo-cp/_rep_policy.xml");
-            notExpected.add("jcr_root/home/users/system/.content.xml");
-            notExpected.add("jcr_root/home/users/system/_cq_services/.content.xml");
-            notExpected.add("jcr_root/home/users/system/_cq_services/demo-cp/.content.xml");
-            notExpected.add("jcr_root/home/users/system/_cq_services/demo-cp/qStDu7IQBLa95gURmer1/.content.xml");
-            notExpected.add("jcr_root/home/users/system/_cq_services/demo-cp/qStDu7IQBLa95gURmer1/_rep_principalPolicy.xml");
-            verifyContentPackage(converted,
-                    notExpected,
-                    "META-INF/vault/properties.xml",
-                    "META-INF/vault/config.xml",
-                    "META-INF/vault/filter.xml",
-                    "jcr_root/.content.xml",
-                    "jcr_root/demo-cp/.content.xml",
-                    "jcr_root/demo-cp/_rep_policy.xml",
-                    "jcr_root/apps/.content.xml",
-                    "jcr_root/home/.content.xml",
-                    "jcr_root/home/users/demo-cp/.content.xml",
-                    "jcr_root/home/users/demo-cp/XPXhA_RKMFRKNO8ViIhn/.content.xml",
-                    "jcr_root/home/users/demo-cp/XPXhA_RKMFRKNO8ViIhn/_rep_policy.xml",
-                    "jcr_root/home/groups/.content.xml",
-                    "jcr_root/home/groups/demo-cp/.content.xml",
-                    "jcr_root/home/groups/demo-cp/EsYrXeBdSRkna2kqbxjl/.content.xml",
-                    "jcr_root/home/groups/demo-cp/EsYrXeBdSRkna2kqbxjl/_rep_policy.xml"
-            );
-
-            assertPolicy(converted, "jcr_root/demo-cp/_rep_policy.xml", "cp-serviceuser-1", "cp-user1", "cp-group1");
-            assertPolicy(converted, "jcr_root/home/groups/demo-cp/EsYrXeBdSRkna2kqbxjl/_rep_policy.xml", null, "cp-group1");
-            assertPolicy(converted,  "jcr_root/home/users/demo-cp/XPXhA_RKMFRKNO8ViIhn/_rep_policy.xml", null, "cp-user1");
-        } finally {
-            deleteDirTree(outputDirectory);
-        }
-    }
-    @Test
-    public void testConvertCONTENTPackageWithUsersGroupsAndServiceUsers() throws Exception {
-        URL packageUrl = getClass().getResource("demo-cp2.zip");
-        File packageFile = FileUtils.toFile(packageUrl);
-        File outputDirectory = new File(System.getProperty("java.io.tmpdir"), getClass().getName() + '_' + System.currentTimeMillis());
-        File unrefOutputDir = new File(outputDirectory, "unref");
-        try {
-            converter.setFeaturesManager(new DefaultFeaturesManager(true, 5, outputDirectory, null, null, null, new DefaultAclManager()))
-                    .setBundlesDeployer(new LocalMavenRepositoryArtifactsDeployer(outputDirectory))
-                    .setUnreferencedArtifactsDeployer(new LocalMavenRepositoryArtifactsDeployer(unrefOutputDir))
-                    .setContentTypePackagePolicy(PackagePolicy.PUT_IN_DEDICATED_FOLDER)
-                    .setEmitter(DefaultPackagesEventsEmitter.open(outputDirectory))
-                    .convert(packageFile);
-
-            File converted = new File(unrefOutputDir, "my_packages/demo-cp/0.0.0/demo-cp-0.0.0-cp2fm-converted.zip");
-            Set<String> notExpected = new HashSet<>();
-            notExpected.add("jcr_root/home/users/demo-cp/_rep_policy.xml");
-            notExpected.add("jcr_root/home/groups/demo-cp/_rep_policy.xml");
-            notExpected.add("jcr_root/home/users/system/.content.xml");
-            notExpected.add("jcr_root/home/users/system/_cq_services/.content.xml");
-            notExpected.add("jcr_root/home/users/system/_cq_services/demo-cp/.content.xml");
-            notExpected.add("jcr_root/home/users/system/_cq_services/demo-cp/qStDu7IQBLa95gURmer1/.content.xml");
-            notExpected.add("jcr_root/home/users/system/_cq_services/demo-cp/qStDu7IQBLa95gURmer1/_rep_principalPolicy.xml");
-            verifyContentPackage(converted,
-                    notExpected,
-                    "META-INF/vault/properties.xml",
-                    "META-INF/vault/config.xml",
-                    "META-INF/vault/filter.xml",
-                    "jcr_root/.content.xml",
-                    "jcr_root/demo-cp/.content.xml",
-                    "jcr_root/demo-cp/_rep_policy.xml",
-                    "jcr_root/home/.content.xml",
-                    "jcr_root/home/users/demo-cp/.content.xml",
-                    "jcr_root/home/users/demo-cp/XPXhA_RKMFRKNO8ViIhn/.content.xml",
-                    "jcr_root/home/users/demo-cp/XPXhA_RKMFRKNO8ViIhn/_rep_policy.xml",
-                    "jcr_root/home/groups/.content.xml",
-                    "jcr_root/home/groups/demo-cp/.content.xml",
-                    "jcr_root/home/groups/demo-cp/EsYrXeBdSRkna2kqbxjl/.content.xml",
-                    "jcr_root/home/groups/demo-cp/EsYrXeBdSRkna2kqbxjl/_rep_policy.xml"
-            );
-
-            assertPolicy(converted, "jcr_root/demo-cp/_rep_policy.xml", "cp-serviceuser-1", "cp-user1", "cp-group1");
-            assertPolicy(converted, "jcr_root/home/groups/demo-cp/EsYrXeBdSRkna2kqbxjl/_rep_policy.xml", null, "cp-group1");
-            assertPolicy(converted,  "jcr_root/home/users/demo-cp/XPXhA_RKMFRKNO8ViIhn/_rep_policy.xml", null, "cp-user1");
-        } finally {
-            deleteDirTree(outputDirectory);
-        }
-    }
-    
-    private static void assertPolicy(@NotNull File contentPackage, @NotNull String path, @Nullable String unExpectedPrincipalName, @NotNull String... expectedPrincipalNames) throws IOException {
-        try (ZipFile zipFile = new ZipFile(contentPackage)) {
-            ZipEntry entry = zipFile.getEntry(path);
-            assertNotNull(entry);
-            assertFalse(entry.isDirectory());
-            
-            try (InputStream in = zipFile.getInputStream(entry)) {
-                String policy = IOUtils.toString(in, StandardCharsets.UTF_8);
-                for (String principalName : expectedPrincipalNames) {
-                    assertTrue(policy.contains("rep:principalName=\""+principalName+"\""));
-                }
-                if (unExpectedPrincipalName != null) {
-                    assertFalse(policy.contains("rep:principalName=\""+unExpectedPrincipalName+"\""));
-                }
-            }
-        } 
-    }
-    
 }
